Fix freelook being replaced with a Java exception
When running `task-7-identify-life-stages-1`, focusing on an inanimate
object would cause `freelook`, `inventory`, and `observation` to be
replaced with `"java.util.NoSuchElementException: None.get"` instead of
the proper value. This was caused by not checking whether an `Option`
variable was set before trying to retrieve one of its variables. It only
happened as part of console output, so it was safe to make the `println`
statement conditional on whether the variable was set.