Finish processing a turn before requesting user input
User input was being requested in the middle of the main game loop,
which caused incorrect data to be logged afterwards. The history for a
turn was being stored with the user input for the next turn. This was
confusing for me when I was trying to analyze the data.