Fix StackOverflowError when a wire is connected to itself
During each tick, electrical components are checked for whether they
connect to ground with the help of the method
`Terminal.connectsToGround()`. The method contains a `maxSteps`
parameter that is decremented on each recursive call to prevent infinite
recursion. However, the value of `maxSteps` was never checked, so the
method would recurse infinitely until the stack overflowed when one
terminal of a wire was connected to the other terminal. Adding checks to
stop recursing when `maxSteps` reaches 0 prevents the error.