Store library version in only a single place
As of now, the library version must be manually set in three different
locations:

- `version` in `simulator/build.sbt`
- `__version__` in `scienceworld/version.py`
- The filename of `scienceworld/scienceworld-*.jar`

This has repeatedly been shown to be problematic because the version
must be changed in all places at once. This commit now uses the version
in `build.sbt` as the single source of truth. `scienceworld.jar` does
not need to be renamed for each version updated, as the file contains
the version in its metadata. `version.py` is set on installation using
the JAR file's metadata. Going forward, only one copy of the version
will need to be manually changed.